Rename package dir to orchestrate
