+++ proteus-orchestrate/cmd/start.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/thetorproject/proteus/proteus-orchestrate/events"
+	"github.com/thetorproject/proteus/proteus-orchestrate/orchestrate"
 )
 
 // startCmd represents the start command
